isolate the resource rendering for further reuse
