+++ app/controllers/api/v1/base_controller.rb
@@ -4,6 +4,7 @@ class Api::V1::BaseController < ActionController::Base
   skip_before_action :verify_authenticity_token
   include Pundit
   include DeviseTokenAuth::Concerns::SetUserByToken
+  include ApiResourcesRenderer
   respond_to :json
   before_action :authenticate_api_v1_agent_with_token_auth!
 
@@ -26,36 +27,6 @@ class Api::V1::BaseController < ActionController::Base
 
   def authorize(record, *args)
     super([:agent, record], *args)
-  end
-
-  def params
-    params = super
-    @page ||= params.delete(:page)&.to_i || 1
-    @per ||= params.delete(:per)&.to_i || 100
-    params
-  end
-
-  def render_record(record, **options)
-    record_klass = record.class
-    blueprint_klass = "#{record_klass.name}Blueprint".constantize
-    root = record.class.model_name.element
-    render json: blueprint_klass.render(record, root: root, **options)
-  end
-
-  def render_collection(objects)
-    objects = objects.page(@page).per(@per)
-    meta = {
-      current_page: objects.current_page,
-      next_page: objects.next_page,
-      prev_page: objects.prev_page,
-      total_pages: objects.total_pages,
-      total_count: objects.total_count,
-    }
-
-    objects_klass = objects.klass
-    blueprint_klass = "#{objects_klass.name}Blueprint".constantize
-    root = objects_klass.model_name.collection
-    render json: blueprint_klass.render(objects, root: root, meta: meta)
   end
 
   # Rescuable exceptions
